# Improve test performance