--- a/v2/api/authorization/v1alpha1api20200801previewstorage/role_assignment_types_gen_test.go
+++ b/v2/api/authorization/v1alpha1api20200801previewstorage/role_assignment_types_gen_test.go
@@ -22,6 +22,7 @@ func Test_RoleAssignment_WhenConvertedToHub_RoundTripsWithoutLoss(t *testing.T) 
 	t.Parallel()
 	parameters := gopter.DefaultTestParameters()
 	parameters.MaxSize = 10
+	parameters.MinSuccessfulTests = 10
 	properties := gopter.NewProperties(parameters)
 	properties.Property(
 		"Round trip from RoleAssignment to hub returns original",