--- a/v2/api/batch/v1alpha1api20210101storage/batch_account_types_gen_test.go
+++ b/v2/api/batch/v1alpha1api20210101storage/batch_account_types_gen_test.go
@@ -22,6 +22,7 @@ func Test_BatchAccount_WhenConvertedToHub_RoundTripsWithoutLoss(t *testing.T) {
 	t.Parallel()
 	parameters := gopter.DefaultTestParameters()
 	parameters.MaxSize = 10
+	parameters.MinSuccessfulTests = 10
 	properties := gopter.NewProperties(parameters)
 	properties.Property(
 		"Round trip from BatchAccount to hub returns original",