--- a/v2/api/cache/v1alpha1api20201201/redis_linked_server_types_gen_test.go
+++ b/v2/api/cache/v1alpha1api20201201/redis_linked_server_types_gen_test.go
@@ -23,6 +23,7 @@ func Test_RedisLinkedServer_WhenConvertedToHub_RoundTripsWithoutLoss(t *testing.
 	t.Parallel()
 	parameters := gopter.DefaultTestParameters()
 	parameters.MaxSize = 10
+	parameters.MinSuccessfulTests = 10
 	properties := gopter.NewProperties(parameters)
 	properties.Property(
 		"Round trip from RedisLinkedServer to hub returns original",