--- a/v2/api/cache/v1alpha1api20201201/redis_patch_schedule_types_gen_test.go
+++ b/v2/api/cache/v1alpha1api20201201/redis_patch_schedule_types_gen_test.go
@@ -23,6 +23,7 @@ func Test_RedisPatchSchedule_WhenConvertedToHub_RoundTripsWithoutLoss(t *testing
 	t.Parallel()
 	parameters := gopter.DefaultTestParameters()
 	parameters.MaxSize = 10
+	parameters.MinSuccessfulTests = 10
 	properties := gopter.NewProperties(parameters)
 	properties.Property(
 		"Round trip from RedisPatchSchedule to hub returns original",