--- a/v2/api/cache/v1alpha1api20201201storage/redis_firewall_rule_types_gen_test.go
+++ b/v2/api/cache/v1alpha1api20201201storage/redis_firewall_rule_types_gen_test.go
@@ -22,6 +22,7 @@ func Test_RedisFirewallRule_WhenConvertedToHub_RoundTripsWithoutLoss(t *testing.
 	t.Parallel()
 	parameters := gopter.DefaultTestParameters()
 	parameters.MaxSize = 10
+	parameters.MinSuccessfulTests = 10
 	properties := gopter.NewProperties(parameters)
 	properties.Property(
 		"Round trip from RedisFirewallRule to hub returns original",