--- a/v2/api/cache/v1alpha1api20201201storage/redis_types_gen_test.go
+++ b/v2/api/cache/v1alpha1api20201201storage/redis_types_gen_test.go
@@ -23,6 +23,7 @@ func Test_Redis_WhenConvertedToHub_RoundTripsWithoutLoss(t *testing.T) {
 	t.Parallel()
 	parameters := gopter.DefaultTestParameters()
 	parameters.MaxSize = 10
+	parameters.MinSuccessfulTests = 10
 	properties := gopter.NewProperties(parameters)
 	properties.Property(
 		"Round trip from Redis to hub returns original",