--- a/v2/api/cache/v1beta20201201/redis_types_gen_test.go
+++ b/v2/api/cache/v1beta20201201/redis_types_gen_test.go
@@ -22,6 +22,7 @@ func Test_Redis_WhenConvertedToHub_RoundTripsWithoutLoss(t *testing.T) {
 	t.Parallel()
 	parameters := gopter.DefaultTestParameters()
 	parameters.MaxSize = 10
+	parameters.MinSuccessfulTests = 10
 	properties := gopter.NewProperties(parameters)
 	properties.Property(
 		"Round trip from Redis to hub returns original",