--- a/v2/api/cache/v1beta20210301/redis_enterprise_database_types_gen_test.go
+++ b/v2/api/cache/v1beta20210301/redis_enterprise_database_types_gen_test.go
@@ -22,6 +22,7 @@ func Test_RedisEnterpriseDatabase_WhenConvertedToHub_RoundTripsWithoutLoss(t *te
 	t.Parallel()
 	parameters := gopter.DefaultTestParameters()
 	parameters.MaxSize = 10
+	parameters.MinSuccessfulTests = 10
 	properties := gopter.NewProperties(parameters)
 	properties.Property(
 		"Round trip from RedisEnterpriseDatabase to hub returns original",