--- a/v2/api/cdn/v1beta20210601/profile_types_gen_test.go
+++ b/v2/api/cdn/v1beta20210601/profile_types_gen_test.go
@@ -22,6 +22,7 @@ func Test_Profile_WhenConvertedToHub_RoundTripsWithoutLoss(t *testing.T) {
 	t.Parallel()
 	parameters := gopter.DefaultTestParameters()
 	parameters.MaxSize = 10
+	parameters.MinSuccessfulTests = 10
 	properties := gopter.NewProperties(parameters)
 	properties.Property(
 		"Round trip from Profile to hub returns original",