--- a/v2/api/compute/v1alpha1api20200930storage/snapshot_types_gen_test.go
+++ b/v2/api/compute/v1alpha1api20200930storage/snapshot_types_gen_test.go
@@ -22,6 +22,7 @@ func Test_Snapshot_WhenConvertedToHub_RoundTripsWithoutLoss(t *testing.T) {
 	t.Parallel()
 	parameters := gopter.DefaultTestParameters()
 	parameters.MaxSize = 10
+	parameters.MinSuccessfulTests = 10
 	properties := gopter.NewProperties(parameters)
 	properties.Property(
 		"Round trip from Snapshot to hub returns original",