--- a/v2/api/compute/v1alpha1api20201201/virtual_machine_types_gen_test.go
+++ b/v2/api/compute/v1alpha1api20201201/virtual_machine_types_gen_test.go
@@ -23,6 +23,7 @@ func Test_VirtualMachine_WhenConvertedToHub_RoundTripsWithoutLoss(t *testing.T) 
 	t.Parallel()
 	parameters := gopter.DefaultTestParameters()
 	parameters.MaxSize = 10
+	parameters.MinSuccessfulTests = 10
 	properties := gopter.NewProperties(parameters)
 	properties.Property(
 		"Round trip from VirtualMachine to hub returns original",