--- a/v2/api/compute/v1beta20210701/image_types_gen_test.go
+++ b/v2/api/compute/v1beta20210701/image_types_gen_test.go
@@ -22,6 +22,7 @@ func Test_Image_WhenConvertedToHub_RoundTripsWithoutLoss(t *testing.T) {
 	t.Parallel()
 	parameters := gopter.DefaultTestParameters()
 	parameters.MaxSize = 10
+	parameters.MinSuccessfulTests = 10
 	properties := gopter.NewProperties(parameters)
 	properties.Property(
 		"Round trip from Image to hub returns original",