--- a/v2/api/containerinstance/v1beta20211001/container_group_types_gen_test.go
+++ b/v2/api/containerinstance/v1beta20211001/container_group_types_gen_test.go
@@ -22,6 +22,7 @@ func Test_ContainerGroup_WhenConvertedToHub_RoundTripsWithoutLoss(t *testing.T) 
 	t.Parallel()
 	parameters := gopter.DefaultTestParameters()
 	parameters.MaxSize = 10
+	parameters.MinSuccessfulTests = 10
 	properties := gopter.NewProperties(parameters)
 	properties.Property(
 		"Round trip from ContainerGroup to hub returns original",