--- a/v2/api/containerregistry/v1beta20210901/registry_types_gen_test.go
+++ b/v2/api/containerregistry/v1beta20210901/registry_types_gen_test.go
@@ -22,6 +22,7 @@ func Test_Registry_WhenConvertedToHub_RoundTripsWithoutLoss(t *testing.T) {
 	t.Parallel()
 	parameters := gopter.DefaultTestParameters()
 	parameters.MaxSize = 10
+	parameters.MinSuccessfulTests = 10
 	properties := gopter.NewProperties(parameters)
 	properties.Property(
 		"Round trip from Registry to hub returns original",