--- a/v2/api/containerservice/v1beta20210501/managed_cluster_types_gen_test.go
+++ b/v2/api/containerservice/v1beta20210501/managed_cluster_types_gen_test.go
@@ -22,6 +22,7 @@ func Test_ManagedCluster_WhenConvertedToHub_RoundTripsWithoutLoss(t *testing.T) 
 	t.Parallel()
 	parameters := gopter.DefaultTestParameters()
 	parameters.MaxSize = 10
+	parameters.MinSuccessfulTests = 10
 	properties := gopter.NewProperties(parameters)
 	properties.Property(
 		"Round trip from ManagedCluster to hub returns original",