--- a/v2/api/containerservice/v1beta20210501/managed_clusters_agent_pool_types_gen_test.go
+++ b/v2/api/containerservice/v1beta20210501/managed_clusters_agent_pool_types_gen_test.go
@@ -22,6 +22,7 @@ func Test_ManagedClustersAgentPool_WhenConvertedToHub_RoundTripsWithoutLoss(t *t
 	t.Parallel()
 	parameters := gopter.DefaultTestParameters()
 	parameters.MaxSize = 10
+	parameters.MinSuccessfulTests = 10
 	properties := gopter.NewProperties(parameters)
 	properties.Property(
 		"Round trip from ManagedClustersAgentPool to hub returns original",