--- a/v2/api/dbformariadb/v1beta20180601/configuration_types_gen_test.go
+++ b/v2/api/dbformariadb/v1beta20180601/configuration_types_gen_test.go
@@ -22,6 +22,7 @@ func Test_Configuration_WhenConvertedToHub_RoundTripsWithoutLoss(t *testing.T) {
 	t.Parallel()
 	parameters := gopter.DefaultTestParameters()
 	parameters.MaxSize = 10
+	parameters.MinSuccessfulTests = 10
 	properties := gopter.NewProperties(parameters)
 	properties.Property(
 		"Round trip from Configuration to hub returns original",