--- a/v2/api/dbformysql/v1alpha1api20210501storage/flexible_servers_firewall_rule_types_gen_test.go
+++ b/v2/api/dbformysql/v1alpha1api20210501storage/flexible_servers_firewall_rule_types_gen_test.go
@@ -22,6 +22,7 @@ func Test_FlexibleServersFirewallRule_WhenConvertedToHub_RoundTripsWithoutLoss(t
 	t.Parallel()
 	parameters := gopter.DefaultTestParameters()
 	parameters.MaxSize = 10
+	parameters.MinSuccessfulTests = 10
 	properties := gopter.NewProperties(parameters)
 	properties.Property(
 		"Round trip from FlexibleServersFirewallRule to hub returns original",