--- a/v2/api/dbformysql/v1beta20210501/flexible_server_types_gen_test.go
+++ b/v2/api/dbformysql/v1beta20210501/flexible_server_types_gen_test.go
@@ -22,6 +22,7 @@ func Test_FlexibleServer_WhenConvertedToHub_RoundTripsWithoutLoss(t *testing.T) 
 	t.Parallel()
 	parameters := gopter.DefaultTestParameters()
 	parameters.MaxSize = 10
+	parameters.MinSuccessfulTests = 10
 	properties := gopter.NewProperties(parameters)
 	properties.Property(
 		"Round trip from FlexibleServer to hub returns original",