--- a/v2/api/dbformysql/v1beta20210501/flexible_servers_database_types_gen_test.go
+++ b/v2/api/dbformysql/v1beta20210501/flexible_servers_database_types_gen_test.go
@@ -22,6 +22,7 @@ func Test_FlexibleServersDatabase_WhenConvertedToHub_RoundTripsWithoutLoss(t *te
 	t.Parallel()
 	parameters := gopter.DefaultTestParameters()
 	parameters.MaxSize = 10
+	parameters.MinSuccessfulTests = 10
 	properties := gopter.NewProperties(parameters)
 	properties.Property(
 		"Round trip from FlexibleServersDatabase to hub returns original",