--- a/v2/api/dbforpostgresql/v1alpha1api20210601/flexible_servers_firewall_rule_types_gen_test.go
+++ b/v2/api/dbforpostgresql/v1alpha1api20210601/flexible_servers_firewall_rule_types_gen_test.go
@@ -23,6 +23,7 @@ func Test_FlexibleServersFirewallRule_WhenConvertedToHub_RoundTripsWithoutLoss(t
 	t.Parallel()
 	parameters := gopter.DefaultTestParameters()
 	parameters.MaxSize = 10
+	parameters.MinSuccessfulTests = 10
 	properties := gopter.NewProperties(parameters)
 	properties.Property(
 		"Round trip from FlexibleServersFirewallRule to hub returns original",