--- a/v2/api/documentdb/v1alpha1api20210515/mongodb_database_collection_throughput_setting_types_gen_test.go
+++ b/v2/api/documentdb/v1alpha1api20210515/mongodb_database_collection_throughput_setting_types_gen_test.go
@@ -23,6 +23,7 @@ func Test_MongodbDatabaseCollectionThroughputSetting_WhenConvertedToHub_RoundTri
 	t.Parallel()
 	parameters := gopter.DefaultTestParameters()
 	parameters.MaxSize = 10
+	parameters.MinSuccessfulTests = 10
 	properties := gopter.NewProperties(parameters)
 	properties.Property(
 		"Round trip from MongodbDatabaseCollectionThroughputSetting to hub returns original",