--- a/v2/api/documentdb/v1alpha1api20210515/mongodb_database_collection_types_gen_test.go
+++ b/v2/api/documentdb/v1alpha1api20210515/mongodb_database_collection_types_gen_test.go
@@ -23,6 +23,7 @@ func Test_MongodbDatabaseCollection_WhenConvertedToHub_RoundTripsWithoutLoss(t *
 	t.Parallel()
 	parameters := gopter.DefaultTestParameters()
 	parameters.MaxSize = 10
+	parameters.MinSuccessfulTests = 10
 	properties := gopter.NewProperties(parameters)
 	properties.Property(
 		"Round trip from MongodbDatabaseCollection to hub returns original",