… v2/api/documentdb/v1alpha1api20210515/mongodb_database_throughput_setting_types_gen_test.go
… v2/api/documentdb/v1alpha1api20210515/mongodb_database_types_gen_test.go
… v2/api/documentdb/v1alpha1api20210515/sql_database_container_stored_procedure_types_gen_test.go
… v2/api/documentdb/v1alpha1api20210515/sql_database_container_throughput_setting_types_gen_test.go
… v2/api/documentdb/v1alpha1api20210515/sql_database_container_trigger_types_gen_test.go
… v2/api/documentdb/v1alpha1api20210515/sql_database_container_types_gen_test.go
… v2/api/documentdb/v1alpha1api20210515/sql_database_container_user_defined_function_types_gen_test.go
… v2/api/documentdb/v1alpha1api20210515/sql_database_throughput_setting_types_gen_test.go
… v2/api/documentdb/v1alpha1api20210515/sql_database_types_gen_test.go
… v2/api/documentdb/v1alpha1api20210515storage/database_account_types_gen_test.go
… v2/api/documentdb/v1alpha1api20210515storage/mongodb_database_collection_throughput_setting_types_gen_test.go
… v2/api/documentdb/v1alpha1api20210515storage/mongodb_database_collection_types_gen_test.go
… v2/api/documentdb/v1alpha1api20210515storage/mongodb_database_throughput_setting_types_gen_test.go
… v2/api/documentdb/v1alpha1api20210515storage/mongodb_database_types_gen_test.go
… v2/api/documentdb/v1alpha1api20210515storage/sql_database_container_stored_procedure_types_gen_test.go
… v2/api/documentdb/v1alpha1api20210515storage/sql_database_container_throughput_setting_types_gen_test.go
… v2/api/documentdb/v1alpha1api20210515storage/sql_database_container_trigger_types_gen_test.go
… v2/api/documentdb/v1alpha1api20210515storage/sql_database_container_types_gen_test.go
… v2/api/documentdb/v1alpha1api20210515storage/sql_database_container_user_defined_function_types_gen_test.go
… v2/api/documentdb/v1alpha1api20210515storage/sql_database_throughput_setting_types_gen_test.go
… v2/api/documentdb/v1alpha1api20210515storage/sql_database_types_gen_test.go
… v2/api/documentdb/v1beta20210515/database_account_types_gen_test.go
… v2/api/documentdb/v1beta20210515/mongodb_database_collection_throughput_setting_types_gen_test.go
… v2/api/documentdb/v1beta20210515/mongodb_database_collection_types_gen_test.go
… v2/api/documentdb/v1beta20210515/mongodb_database_throughput_setting_types_gen_test.go
… v2/api/documentdb/v1beta20210515/mongodb_database_types_gen_test.go
… v2/api/documentdb/v1beta20210515/sql_database_container_stored_procedure_types_gen_test.go
… v2/api/documentdb/v1beta20210515/sql_database_container_throughput_setting_types_gen_test.go
… v2/api/documentdb/v1beta20210515/sql_database_container_trigger_types_gen_test.go
… v2/api/documentdb/v1beta20210515/sql_database_container_types_gen_test.go