**What this PR does / why we need it**:

Our tests are taking too long to run, so we need to reduce the time spent.

We have separate tests that check each step of the generated resource conversion graph works in isolation; we also have tests that check they work in concert together, conducting round trip checks to the hub type for each resource.

Given we have good property-testing coverage (via gopter) of the individual steps, it seems safe to reduce the number of tests we run of the entire round trip.

Performance generally looks better (results from a single run only, so take with a grain of salt):

| Package                                |  Before |    After |
| -------------------------------------- | ------: | -------: |
| api/compute/v1alpha1api20200930        | 19.768s |   9.263s |
| api/compute/v1alpha1api20200930storage | 16.145s |  12.331s |
| api/compute/v1alpha1api20201201        | 35.509s |  27.243s |
| api/compute/v1alpha1api20201201storage | 35.017s |  32.165s |
| api/compute/v1alpha1api20210701        | 14.404s |   6.937s |
| api/compute/v1alpha1api20210701storage | 15.339s |   6.718s |
| api/compute/v1beta20200930             | 19.094s |  15.635s |
| api/compute/v1beta20201201             | 26.636s |  26.532s |
| api/compute/v1beta20201201storage      | 28.967s |  29.254s |
| api/compute/v1beta20210701             |  9.147s |  16.722s |
| api/compute/v1beta20210701storage      | 17.605s |  14.032s |
| api/compute/v1beta20220301             | 18.459s |  14.606s |

**Special notes for your reviewer**:

Changes to generated files are the addition of one configuration assignment to each test:

``` golang
parameters.MinSuccessfulTests = 10
```

**How does this PR make you feel**:
![gif](https://media.giphy.com/media/J6VR03Q4GRO12/giphy.gif)

**If applicable**:
- [x] this PR contains tests