@theunrepentantgeek 2022-07-24
`MaxSize` is the maximum size of any generated map or slice.
`MinSuccessfulTests` is the minimum count of successful tests required before the test run stops. 

They're orthogonal concerns.